backend/requirements.txt
python-dotenv==1.0.0
supabase==1.0.3
openai==1.12.0
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
requests==2.31.0
flask==3.0.2
flask-cors==4.0.0 